==> analysis/02_sites.Rmd <==
---
title: "Sites"
output:
  html_notebook:
    css: style.css
    theme: paper
---

```{r setup, message=FALSE}
library(tidyverse)
library(cowplot)
library(maps)
library(viridis)
```

```{r, message=FALSE}
studies <- read_csv("../data/studies_gsheet.csv")
sites <- read_csv("../data/sites_gsheet.csv")
```

```{r}
df <- studies %>% 
  group_by(site) %>% 
  summarise(Nstudies = n_distinct(studyID),
            Nspecies = n_distinct(species)) %>% 
  mutate(legend_species = ifelse(Nspecies > 4, "more      ", Nspecies))%>%
  left_join(sites)
```

```{r}
df %>% 
  arrange(desc(Nspecies), desc(Nstudies)) %>% 
  mutate(site2 = str_sub(site, 1, 40)) %>% 
  select(site2, country, Nstudies, Nspecies)
```

```{r}
world <- map("world", plot = FALSE, fill = TRUE) %>% fortify()
```

```{r, fig.width=20, fig.height=10, cache=TRUE, eval=FALSE}
# sanity check that coordinates are in the right country
# within a country, all Xs should be the same color
ggplot(df, aes(x = long, y = lat)) +
  geom_polygon(data = world, aes(group = group), col = "black", fill = "grey90") +
  geom_point(aes(col = country), size = 5, shape = 4, stroke = 2, alpha = .8) +
  coord_fixed(1.3, ylim = c(-55, 83.59961)) +
  scale_color_viridis_d() +
  theme_map() +
  guides(col = "none")

ggsave("../graphs/map.pdf", width = 20, height = 10, scale = 2)
```

```{r, fig.width=8, fig.height=4}
ms <- ggplot(df, aes(x = long, y = lat)) +
  geom_polygon(data = world, aes(group = group), fill = "grey90") +
  geom_point(aes(size = Nstudies, fill = legend_species), shape = 21, alpha = .8) +
  theme_map() +
  # to have them side by side
  scale_fill_brewer("Species", palette = "Greens") +
  scale_size_area("Studies", max_size = 15, breaks = c(1, 20, 40, 60)) +
  theme(legend.box = "horizontal") +
  guides(fill = guide_legend(override.aes = list(size = 6), keyheight = 1.6, order = 1),
         size = guide_legend(keyheight = 1.6, order = 0))

legend <- get_legend(ms)
# plot_grid(legend)

msx <- ms +
  annotate("rect", ymin = c(24, 31, 24), ymax = c(46, 62, 50), xmin = c(124, -12, -130), 
           xmax = c(146, 32, -64), size = .5, col = "grey", fill = NA) + 
  coord_fixed(1.3, ylim = c(-55, 83.59961)) +
  theme(legend.position = "none")
```

```{r}
# japan
ms1 <- ms + 
  geom_polygon(data = world, aes(group = group), col = "black", fill = "grey90", lwd = .2) +
  geom_point(aes(size = Nstudies, fill = legend_species), shape = 21, alpha = .8) +
  coord_fixed(1.3, ylim = c(25, 45), xlim = c(125, 145)) +
  theme(legend.position = "none") +
  panel_border(color = "grey", size = .5)

# europe
ms2 <- ms + 
  geom_polygon(data = world, aes(group = group), col = "black", fill = "grey90", lwd = .2) +
  geom_point(aes(size = Nstudies, fill = legend_species), shape = 21, alpha = .8) +
  coord_fixed(1.3, ylim = c(32, 60), xlim = c(-10, 30)) +
  theme(legend.position = "none") +
  panel_border(color = "grey", size = .5)

# contiguous us
usa = map("state", plot = FALSE, fill = TRUE) %>% fortify() # "usa"

ms3 <- ms + 
  geom_polygon(data = world, aes(group = group), col = "black", fill = "grey90", lwd = .2) +
  geom_polygon(data = usa, aes(group = group), col = "black", fill = "grey90", lwd = .2) +
  geom_point(aes(size = Nstudies, fill = legend_species), shape = 21, alpha = .8) +
  coord_fixed(1.3, ylim = c(25, 49), xlim = c(-127, -67)) +
  theme(legend.position = "none") +
  panel_border(color = "grey", size = .5)
```

```{r, fig.width=8, fig.height=5.5, cache=TRUE}
# make bottom row (3 map insets + legend)
bottom_row <- plot_grid(ms3, ms2, NA, ms1, NA, legend, nrow = 1, label_x = -.27, label_y = .93,
                        labels = c(rep("", 5), "Number of ..."), 
                        rel_widths = c(2, 1.18, .06, .85, .08, .7))

# make whole plot as 1 column, big map on top, bottom row underneath
plot_grid(msx, bottom_row, ncol = 1, rel_heights = c(2.4, 1)) +
  theme(plot.margin = unit(c(.5, .5, 1.5, .5), "cm"))
```

```{r}
ggsave("../graphs/map_insets.pdf", width = 8, height = 5.5, scale = 2)
ggsave("../graphs/map_insets.png", width = 8, height = 5.5, scale = 2)
ggsave("../graphs/map_insets.tiff", width = 8, height = 5.5, scale = 2, type = "cairo", 
       compression = "lzw")
```


***

# Session Info

```{r}
sessionInfo()
```
