==> analysis/03_phylo_tree.Rmd <==
---
title: "Phylogenetic Tree"
output: 
  html_notebook:
    css: style.css
    theme: paper
---

Load packages

```{r, message=FALSE}
library(tidyverse)
library(gridExtra)
library(cowplot)
library(viridis)
library(ggridges)
library(ggstance)
library(treeio)
library(ggtree)
library(tidytree)
```

Load and combine data

```{r, message=FALSE}
studies <- read_csv("../data/studies_gsheet.csv")
species <- read_csv("../data/species_gsheet.csv")
fulltree <- read.nexus("../data/consensusTree_10kTrees_298Primates_V3.nex")
refs <- read_csv("../data/ref_nodes.csv")
```

```{r}
data <- studies %>% 
  left_join(species, by = c("species" = "species_file")) %>% 
  rename(label = species_tree, label2 = updated_nomenclature, num = n) %>% 
  mutate(label2 = ifelse(is.na(label2), label, label2),
         label2 = fct_recode(label2,
                             "Sapajus_spp." = "Sapajus_apella",
                             "Hoolock_leuconedys*" = "Hoolock_hoolock",
                             "Plecturocebus_cupreus*" = "Plecturocebus_moloch",
                             "Callicebus_nigrifrons*" = "Callicebus_personatus",
                             "Eulemur_rufifrons*" = "Eulemur_rufus")) %>% 
  select(label, label2, studyID, species, site, num)
```

```{r}
data2 <- data %>%
  group_by(label, label2, species) %>% 
  summarise(Nsites = n_distinct(site), Nstudies = n_distinct(studyID)) %>% 
  mutate(div_score = Nsites/Nstudies)
```

```{r}
species <- species %>% 
  mutate(updated_nomenclature = fct_recode(updated_nomenclature,
                                           "Sapajus_spp." = "Sapajus_apella",
                                           "Hoolock_leuconedys*" = "Hoolock_hoolock",
                                           "Plecturocebus_cupreus*" = "Plecturocebus_moloch",
                                           "Callicebus_nigrifrons*" = "Callicebus_personatus",
                                           "Eulemur_rufifrons*" = "Eulemur_rufus"))
```

```{r}
# turn tree into tidy dataframe
tree2 <- fulltree %>% 
  drop.tip(c("Pan_troglodytes_schweinfurthii", "Pan_troglodytes_troglodytes",
             "Pan_troglodytes_vellerosus", "Pongo_pygmaeus", "Cercopithecus_cephus_cephus",
             "Cercopithecus_erythrogaster_erythrogaster", "Eulemur_fulvus_mayottensis",
             "Hapalemur_griseus_griseus", "Microcebus_lokobensis", "Lepilemur_mitsinjoensis",
             "Gorilla_gorilla_graueri", "Cebus_xanthosternos")) %>%
  as_tibble

inner_nodes <- c(287:296, 305:306, 311:317, 319:323, 325, 327:344, 347:349, 353:355, 357, 373:381, 386, 390:393, 395:401, 403:412, 420:423, 425:429, 435:439, 443:445, 447:451, 454:461, 463:467, 469:473, 477:488, 500, 518:519, 523:525, 528:532, 538:540, 546:548, 551, 553)

tree3 <- tree2 %>% 
  mutate(label = fct_recode(label, 
                            "Pongo_spp." = "Pongo_abelii",
                            "Pan_troglodytes" = "Pan_troglodytes_verus")) %>% 
  left_join(data2) %>% 
  left_join(species, by = c("label" = "species_tree")) %>% 
  select(-label, -label2, -species_file) %>%
  rename(label = updated_nomenclature) %>%
  mutate(
    hasN = ifelse(is.na(Nstudies), .1, .5), # used to size branches + color the tip labels
    hasN2 = ifelse(is.na(Nstudies) & !(node %in% inner_nodes), 0, .5), # used to color branches
    label = str_replace_all(label, "_", " "),
    # label = fct_recode(label, "Sapajus spp." = "Sapajus apella", "Hoolock hoolock*" = "Hoolock hoolock"),
    label = ifelse(is.na(label) & species == "orangutan", "Pongo spp.", 
              ifelse(is.na(label) & species == "chimpanzee", "Pan troglodytes", label))) %>% 
  left_join(refs) %>% 
  groupClade(refs$node[-1]) %>% 
  mutate(group = fct_recode(group, "2" = "1"))

# turn back into tree
tree4 <- as.treedata(tree3)
```

# Circular tree of the 10ktree primates

```{r}
cols <- viridis(4, end = .9)
```

```{r}
p <- ggtree(tree4, aes(alpha = hasN2), layout = "circular") + # size = hasN, 
  # highlight clades with background colors
  geom_hilight(node = 474, fill = cols[1], alpha = .3) +
  geom_hilight(node = 477, fill = cols[1], alpha = .3) +
  geom_hilight(node = 411, fill = cols[2], alpha = .3) +
  geom_hilight(node = 291, fill = cols[3], alpha = .3) +
  geom_hilight(node = 395, fill = cols[4], alpha = .3) +
  # plot tree again to be on top of the highlights
  geom_tree() +
  # root
  geom_rootpoint(size = 1) +
  # tips
  geom_tippoint(aes(size = Nstudies), alpha = .7) +
  geom_tiplab2(aes(alpha = hasN), offset = 3, size = 2.5) +
  # tweak scales
  scale_alpha_continuous(range = c(.2, 1)) +
  scale_size_area(max_size = 8) +
  # widen plotting area
  xlim(NA, 100)

p <- rotate(p, 290)
```

```{r}
pcol <- ggplot(tibble(cols = cols, x = 1:4), aes(x, y = 1, col = cols)) +
  geom_point(size = 6, alpha = .3) +
  scale_color_identity("Clade", guide = "legend", breaks = cols[4:1], 
                       labels = c("Hominoidea", "Cercopithecoidea", "Platyrrhini", 
                                  "Tarsiiformes & Strepsirrhini")) +
  theme_cowplot()

l1 <- get_legend(pcol)
```

```{r}
psize <- ggplot(data2, aes(size = Nstudies, x = 1, y = 1)) +
  geom_point(alpha = .7) +
  scale_size_area("Number of Studies", max_size = 8, breaks = c(1, 5, 10, 25, 50, 100)) +
  theme_cowplot()

l2 <- get_legend(psize)
```

```{r, fig.width=7.5, fig.height=5.5, cache=TRUE}
px <- plot_grid(p, plot_grid(NA, l1, l2, NA, ncol = 1, rel_heights = c(.3, .15, .15, .3)), NA,
          nrow = 1, rel_widths = c(1, .2, .1))

px
```

```{r, cache=TRUE}
ggsave("../graphs/phylo_full.pdf", px, width = 7.5, height = 5.5, scale = 2)
ggsave("../graphs/phylo_full.png", width = 7.5, height = 5.5, scale = 2)
ggsave("../graphs/phylo_full.tiff", width = 7.5, height = 5.5, scale = 2, type = "cairo", 
       compression = "lzw")
```

```{r, fig.width=8, fig.height=8, cache=TRUE, eval=FALSE}
# to figure out node numbers
n1 <- p + geom_text(aes(label = node, x = branch), size = 2, col = "blue", vjust = -.5)
ggsave("../graphs/full_tree_nodes_circular.pdf", n1, width = 8, height = 8, scale = 2)
```

```{r, fig.width=8, fig.height=20, cache=TRUE, eval=FALSE}
n2 <- ggtree(tree4, aes(size = hasN, alpha = hasN2)) +
  # highlight clades with background colors
  geom_hilight(node = 474, fill = cols[1], alpha = .3) +
  geom_hilight(node = 477, fill = cols[1], alpha = .3) +
  geom_hilight(node = 411, fill = cols[2], alpha = .3) +
  geom_hilight(node = 291, fill = cols[3], alpha = .3) +
  geom_hilight(node = 395, fill = cols[4], alpha = .3) +
  # plot tree again to be on top of the highlights
  geom_tree() +
  # root
  geom_rootedge(rootedge = 2) +
  geom_rootpoint(size = 1) +
  # node labels
  geom_text(aes(label = node, x = branch), size = 2, col = "blue", vjust = -.5) +
  # tips
  geom_tippoint(aes(size = Nstudies), alpha = .7) +
  geom_tiplab(aes(alpha = hasN), offset = 1.8, size = 3) +
  # tweak scales
  scale_alpha_continuous(range = c(.2, 1)) +
  scale_size_continuous(range = c(.5, 8)) +
  # widen plotting area
  expand_limits(x = 90) +
  theme_tree2()

ggsave("../graphs/full_tree_nodes.pdf", n2, width = 8, height = 20, scale = 2)
```


# Sample size in detail

```{r}
studies
```

```{r}
# subset tree to just those species who have sample sizes reported, i.e. those who were tested
to_drop <- tree3 %>% filter(is.na(Nstudies)) %>% pull(label)
to_drop2 <- data %>% group_by(label2) %>% filter(all(is.na(num))) %>% pull(label2) %>% str_replace_all('_', ' ')
tree5 <- drop.tip(tree4, c(to_drop, to_drop2))
d3 <- data %>% 
  mutate(label = str_replace_all(label2, "_", " ")) %>% 
  group_by(label, species, studyID) %>% 
  summarise(num = sum(num))
```

```{r}
d3 %>% arrange(desc(num))
```

```{r}
# filter super large samples out for visualization? note in caption
# species with more than X studies can get a density
d3a <- d3 %>% group_by(species) %>% filter(n_distinct(studyID) >= 4, num <= 200)
d3b <- d3 %>% # setdiff(d3, d3a) %>% ## <- to NOT show points for densities
  group_by(species) %>% 
  # create variable num2 is NA if there's only one data point for a species
  # --> those species will only get the vertical crossbar
  mutate(flag = n_distinct(studyID) == 1) %>% 
  ungroup %>% 
  mutate(num2 = ifelse(flag, NA, num)) %>% 
  filter(num <= 200)

# for vertical crossbar = median
d4 <- d3 %>% 
  group_by(label, species) %>% 
  summarise(Mdn = median(num, na.rm = T)) # totalN = sum(num), sitesN = n_distinct(site)

# for vertical line in ridge plot (grand median)
# + hacky way to make horizontal grid lines for right panel only
v <- tibble(reference = c(NA, median(d3$num, na.rm = T)), .panel = c("Tree", "xSample size"))
h <- tibble(reference = c(NA, 1:Ntip(tree5)), .panel = c("Tree", rep("xSample size", Ntip(tree5))))

# for axis labels
ax <- tibble(lab = c("Distance (Millions of years)", "Sample size"), 
             x = c(60, 100), y = -4, .panel = c("Tree", "xSample size"))

# Nsites/studies labels
Nlab <- tibble(lab = c("# Sites", "# Studies"), x = c(125, 136), y = Ntip(tree5) + 1, 
             .panel = "Tree")
```

```{r, cache=TRUE}
# LEFT FACET
q <- ggtree(tree5, aes(col = group)) +
  # root
  geom_rootedge(rootedge = 5) +
  # tip labels
  geom_tippoint(aes(size = Nstudies), shape = 21, fill = "white") +
  geom_tippoint(aes(size = Nsites), stroke = 0, alpha = .8) +
  # geom_tiplab(aes(label = str_c(label, " (", Nsites, "/", Nstudies, ")")), offset = 4, size = 3) +
  geom_tiplab(offset = 4, size = 3) +
  geom_text(aes(label = Nsites), x = 135, hjust = 1, size = 3) +
  geom_text(aes(label = Nstudies), x = 142, hjust = 1, size = 3) +
  # tweak scales
  scale_color_manual(values = c("grey30", cols)) +
  scale_fill_manual(values = cols) +
  scale_size_area(max_size = 8) +
  # display timescale at the bottom
  theme_tree2() +
  xlim_tree(142) +
  xlim_expand(c(0, 175), "xSample size") +
  # # node labels if needed for reference
  # geom_text(aes(label = node, x = branch), size = 2, col = "blue", vjust = -.5) +
  # add axis + Nstudies/sites labels
  geom_text(data = ax, aes(label = lab), col = "black") +
  geom_text(data = Nlab, aes(label = lab), col = "black", size = 2.5) +
  scale_x_continuous(expand = expand_scale(mult = c(0, .01))) +
  scale_y_continuous(limits = c(2, Ntip(tree5)-1), oob = function(x, ...) x) +
  coord_cartesian(clip = "off") +
  # add reference lines (these will show up on right panel of facet_plot only)
  geom_hline(data = h, aes(yintercept = reference), lwd = .2, col = "grey", alpha = .5) +
  geom_vline(data = v, aes(xintercept = reference), lwd = 1.5, col = "grey", alpha = .3) +
  # remove facet strips, expand bottom margin (to make space for x axis labels)
  theme(strip.text = element_blank(), strip.background = element_blank(),
        plot.margin = unit(c(1, 1, 2, 1.5), "cm"), panel.spacing = unit(1, "cm"))

q <- rotate(q, 71)
```

```{r, fig.width=6, fig.height=8, cache=TRUE}
# right-side viz depends on the number of sites per species:
# 1 site = vertical crossbar only
# 2+ sites = points + crossbar at median
# X+ sites = densities (currently, X = 4 just to illustrate)

# dirty hack: x in front of "Sample size" is to have that panel sort to the right (alphabetically) until I figure out why it doesn't just go by order. This cropped up as an issue when I added the dummy point for the x-axis expansion...

# ADD RIGHT FACET
qx <- q %>% 
  # densities for species with enough sites
  facet_plot("xSample size", d3a, geom_density_ridges, 
             aes(x = num, group = label, fill = group, height = ..density..),
             alpha = .5, lwd = .3, scale = .3) %>%
  # vertical crossbar for Mdn
  facet_plot("xSample size", d4, geom_crossbarh, aes(x = Mdn, xmin = Mdn, xmax = Mdn, group = label,
             col = group), alpha = .5, width = .6, fatten = 1.5) %>%
  # vertical mark for individual sites
  facet_plot("xSample size", d3b, geom_jitter, aes(x = num2, group = label), shape = "|", size = 2.5,
             width = .5, height = 0, alpha = .5)
```

```{r}
# add legends
psize <-
  ggplot(data2, aes(x = 1, y = 1)) +
  geom_point(aes(size = Nstudies), col = NA) +
  geom_point(aes(size = Nsites), stroke = 0, alpha = .8) +
  scale_size_area("Number of Sites", max_size = 8, breaks = c(1, 5, 10, 25, 50)) +
  theme_cowplot()

psize2 <-
  ggplot(data2, aes(x = 1, y = 1)) +
  geom_point(aes(size = Nstudies), shape = 21, fill = "white") +
  scale_size_area("\nNumber of Studies", max_size = 8, breaks = c(1, 5, 10, 25, 50, 100)) +
  theme_cowplot()

l2 <- get_legend(psize)
l3 <- get_legend(psize2)
```

```{r, fig.width=8, fig.height=8, cache=TRUE}
qx2 <- plot_grid(qx, plot_grid(NA, l1, l2, l3, NA, ncol = 1, rel_heights = c(.3, .1, .1, .1, .3)), NA,
          nrow = 1, rel_widths = c(1, .2, .1))

qx2
```

```{r}
ggsave("../graphs/phylo_ridge_site.pdf", width = 8, height = 8, scale = 2)
ggsave("../graphs/phylo_ridge_site.png", width = 8, height = 8, scale = 2)
ggsave("../graphs/phylo_ridge_site.tiff", width = 8, height = 8, scale = 2, type = "cairo", 
       compression = "lzw")
```

# Diversity score

```{r, eval=FALSE}
# subset tree to just those species who have sample sizes reported, i.e. those who were tested
to_drop <- tree3 %>% filter(is.na(Nstudies) | Nstudies < 2) %>% pull(label)
tree6 <- drop.tip(tree4, to_drop)
```

```{r, fig.width=4, fig.height=5, cache=TRUE, eval=FALSE}
ggtree(tree6, aes(col = group)) +
  # root
  geom_rootedge(rootedge = 5) +
  # tip labels
  geom_tippoint(aes(size = Nstudies), shape = 21, fill = "white") +
  geom_tippoint(aes(size = Nsites), stroke = 0, alpha = .8) +
  geom_tiplab(offset = 4, size = 3) +
  geom_text(aes(label = Nsites), x = 113, hjust = 1, size = 3) +
  geom_text(aes(label = Nstudies), x = 120, hjust = 1, size = 3) +
  # tweak scales
  scale_color_manual(values = c("grey30", cols)) +
  scale_fill_manual(values = cols) +
  scale_size_area(max_size = 8) +
  # display timescale at the bottom
  theme_tree2() +
  xlim_tree(120) +
  xlab("Distance (Millions of years)")
```

```{r}
# ggsave("../graphs/phylo_div_score.pdf", width = 4, height = 4.5, scale = 2)
```

# Session info

```{r}
sessionInfo()
```
