==> analysis/04_descriptives.Rmd <==
---
title: "Descriptives"
output:
  html_notebook:
    css: style.css
    theme: paper
---

```{r setup, message=FALSE}
library(tidyverse)
```

```{r, message=FALSE}
studies <- read_csv("../data/studies_gsheet.csv") %>% 
  mutate(species = ifelse(species == "Macaca sp.", NA, species))
```

# Total number of sites in sample

```{r}
studies %>% summarise(Nsites = n_distinct(site))
```

# Total number of studies and species in sample

```{r}
studies %>% summarise(Nstudies = n_distinct(studyID), Nspecies = n_distinct(species))
```

# Number of studies by year

```{r}
# not super exciting
studies %>% 
    mutate(year = fct_recode(year, "2019" = "in press")) %>% 
    group_by(year) %>% 
    summarise(Nstudies = n_distinct(studyID)) %>% 
ggplot(aes(year, Nstudies)) +
    geom_bar(stat = "identity", col = "black", fill = "white", width = .5) +
    geom_text(aes(label = Nstudies), vjust = 2) +
    scale_y_continuous(expand=expansion(mult = c(0, .05)))
```

# Number of studies & sites featuring rhesus macaques or chimps

```{r}
studies %>% 
    group_by(species) %>% 
    summarise(Nstudies = n_distinct(studyID), Nsites = n_distinct(site)) %>% 
    filter(species %in% c("rhesus macaque", "chimpanzee")) %>% 
    mutate(diversity = round(Nsites/Nstudies, 2))
```

# Percentiles (number of studies by species)

```{r}
sp_stud <- studies %>% 
    group_by(species) %>% 
    summarise(Nstudies = n_distinct(studyID))
```

```{r}
quantile(sp_stud$Nstudies, seq(.05, 1, .05)) %>% round()
```

e.g., 76% of species appeared in fewer than 6 studies

```{r}
sp_stud %>% 
    ungroup %>% 
    summarise(Nx = sum(Nstudies < 6), n(), percent = round(Nx/n(), 2))
```

# Number of species with only 1 study

```{r}
studies %>% 
    group_by(species) %>% 
    summarise(Nstudies = n_distinct(studyID)) %>% 
    summarise(N1study = sum(Nstudies == 1), n(), percent = round(N1study/n(), 2))
```

# Species comparisons

Number of studies involved more than 1 species

```{r}
sp_2plus <- studies %>% 
  group_by(studyID) %>% 
  summarise(Nspecies = n_distinct(species))

sp_2plus %>% 
  summarise(N2plus = sum(Nspecies > 1), n(), percent = round(N2plus/n(), 2))
```

Number of studies compared species quantitatively

```{r}
studies %>% 
  group_by(species_comparison) %>% 
  summarise(n_distinct(studyID))
```

## Number of studies compared more than 2 species

```{r}
sp_3plus_ID <- sp_2plus %>% filter(Nspecies > 2) %>% with(studyID)

studies %>% 
  filter(studyID %in% sp_3plus_ID) %>% 
  group_by(studyID) %>% 
  summarise(Nspcomp = sum(species_comparison == "yes")) %>% 
  summarise(Nsp3pluscomp = sum(Nspcomp > 0))
```

# Sample size

```{r}
sample_sizes <- studies %>% 
    filter(!is.na(n)) %>% 
    group_by(species, studyID) %>% 
    summarize(N = sum(n))
```

## Percentiles

X% of samples (by species & study) have sample sizes of less than/equal to:

```{r}
quantile(sample_sizes$N, seq(.05, 1, .05)) %>% round()
```

e.g., 74% of samples included less than 16 individuals

```{r}
sample_sizes %>% 
    ungroup %>% 
    summarise(Nx = sum(N < 16), n(), percent = round(Nx/n(), 2))
```

Sample size medians by species

```{r}
sample_sizes2 <- studies %>% 
    filter(!is.na(n)) %>% 
    group_by(species, studyID) %>% 
    summarise(N = sum(n)) %>% 
    summarise(Mdn = median(N))
```

```{r}
quantile(sample_sizes2$Mdn, seq(.05, 1, .05)) %>% round()
```

e.g., 66% of species included less than 10 individuals (median)

```{r}
sample_sizes2 %>% 
    ungroup %>% 
    summarise(Mdnx = sum(Mdn < 10), n(), percent = round(Mdnx/n(), 2))
```
## Median/range

> across species & studies

```{r}
# min, 1st quantile, median, 3rd quantile, max
fivenum(sample_sizes$N)
```

> across studies

```{r}
studies %>% 
    filter(!is.na(n)) %>% 
    group_by(studyID) %>% 
    summarise(N = sum(n)) %>% 
    with(fivenum(N))
```

# Percentage of studies from X sites

```{r}
st_stud <- studies %>% 
  filter(!is.na(site)) %>% 
  group_by(site) %>% 
  summarise(Nstudies = n_distinct(studyID)) %>% 
  arrange(desc(Nstudies))
```

e.g., 5 sites were involved in 217 studies (though other sites may have too)

```{r}
(top_sites <- st_stud$site[1:5])

studies %>% 
  filter(site %in% top_sites) %>% 
  summarise(n_distinct(studyID))
```


# Rate of replication attempts

Number of studies with at least 1 replication attempt (count "yes" and "conceptual" at the moment)

```{r}
studies %>% 
  group_by(studyID) %>%
  summarise(Nrep = sum(attempt_replication != "no", na.rm = T)) %>% 
  summarise(sum(Nrep > 0), n())
```
